"search"

+++ src/main/java/com/brightcoding/app/ws/repositories/ExperienceRepository.java
@@ -5,10 +5,12 @@ import com.brightcoding.app.ws.entities.ExperienceEntity;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
+import java.util.Optional;
 
-public interface ExperienceRepository extends CrudRepository<ExperienceEntity, Long> {
+public interface ExperienceRepository extends CrudRepository<ExperienceEntity, Integer> {
    
 
     ExperienceEntity findByExperienceId(String experienceId);
     public ExperienceEntity save(ExperienceEntity experience);
+	
 }
