+++ src/main/java/com/brightcoding/app/ws/repositories/SkillsRepository.java
@@ -8,7 +8,7 @@ import org.springframework.stereotype.Repository;
 
 import java.util.List;
 @Repository
-public interface SkillsRepository extends CrudRepository<SkillsEntity, Long> {
+public interface SkillsRepository extends CrudRepository<SkillsEntity, Integer> {
   //  List<SkillsEntity> findByCondidat(CondidatEntity currentCondidat);
 
     SkillsEntity findBySkillsId(String skillsId);
